use logging.info correctly
The function doesn't print all arguments in order, instead the first argument is treated as a format String.
Therefore, if a second argument is supplied, but not used in the format String that is the first argument, the function throws an error.
Also fixes commented-out uses, in case someone uncomments them.

see also https://docs.python.org/3/library/logging.html#logging.info

Fixes: c490ec802449 ("Initial Commit")

Signed-off-by: Martin Meilinger <martin.meilinger@st.oth-regensburg.de>