deal with pylint R1728: consider-using-generator
regarding the changes in (Reverse)GreedyClassicalPVC:
If I just delete the brackets, I get the error "Generator expression must be parenthesized if not sole argument",
because the min(...)/max(...) function gets a second argument (the lambda).
Therefore, I put parentheses around it.
This doesn't change the actual length of the code, but should yield a small performance improvement,
since the program doesn't have to go through all of these items twice.

Signed-off-by: Martin Meilinger <martin.meilinger@st.oth-regensburg.de>