Bugfix issue #49 (#117)
* Ignore imaginary terms of problem matrix

If the imaginary part of all matrix elements is zero, the matrix is transformed into a real matrix. If there are imaginary terms != 0, a warning is logged.

* Correct reverse_mapping

Updated the input type from dict to any. In all tests run, the input format was a numpy array. Corrected checking for -1 in the solution string.

* Correct linting errors