Implementation of auto-carrier loading problem
The auto-carrier loading problem is an industry-relevant optimization use case in which the goal is to find an optimal assignment of vehicles on platforms of a truck such that the number of vehicles is maximized while all weight and length constraints and regulations are fulfilled. In this implementation three scenarios have been implemented, ranging from a toy problem with only six cars and few constraints to a full-fetched problem instance close to a real-world example with ten cars and complex parameters and constraints. A more in-depth introduction to the application is provided in [http://dx.doi.org/10.2139/ssrn.4513003](https://dx.doi.org/10.2139/ssrn.4513003).

-----

Co-authored-by: Christian Jaeck [Christian.Jaeck@bmw.de](mailto:Christian.Jaeck@bmw.de)
Co-authored-by: Marvin Erdmann [Marvin.Erdmann@bmw.de](mailto:Marvin.Erdmann@bmw.de)
Co-authored-by: Philipp Ross [Philipp.Ross@bmw.de](mailto:Philipp.Ross@bmw.de)
Co-authored-by: Carlos Riofrio [Carlos.Riofrio@bmwgroup.com](mailto:Carlos.Riofrio@bmwgroup.com)