Implementation of QGAN and some minor updates in QML module (#128)
Quantum Generative Adversarial Networks (QGANs) generate synthetic data sets by an interplay of a generator and a discriminator. The generator learns to generate data samples that resemble the original data set such that the discriminator cannot tell apart the synthetic and the original data. Now, this method is implemented as a part of the generative modeling module in QUARK.
Additional content:
* Add LibraryPennylane
* Refinement of docstrings and typings in QML modules

---------

Co-authored-by: Florian Kiwit <florian.kiwit@bmw.de>
Co-authored-by: Marvin Erdmann <marvin.erdmann@bmw.de>
Co-authored-by: Maximilian Wolf <maximilian.mw.wolf@bmw.de>