Implement a way to visualize solutions
I added the function visualize_solution in Optimization.py that can be overridden by any sub-class where a visualization of the solution should be created. The function is called with the preprocessed solution after it has been validated.
I also provided function implementations for the MIS, PVC, and TSP modules.